requirements.txt
pillow>=9.0.0
pyqt6>=6.0.0
watchdog>=2.0.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.20.0
sqlalchemy>=1.4.0
accelerate>=0.20.0
safetensors>=0.3.0
huggingface-hub>=0.14.0
